+++ src/containerapp/azext_containerapp/custom.py
@@ -10,6 +10,8 @@ from urllib.parse import urlparse
 import json
 import subprocess
 from concurrent.futures import ThreadPoolExecutor
+from ._constants import DOTNET_COMPONENT_RESOURCE_TYPE
+
 
 from azure.cli.core import telemetry as telemetry_core
 from azure.cli.command_modules.containerapp._utils import safe_set, safe_get
@@ -78,6 +80,7 @@ from .containerapp_decorator import ContainerAppPreviewCreateDecorator, Containe
 from .containerapp_env_storage_decorator import ContainerappEnvStorageDecorator
 from .java_component_decorator import JavaComponentDecorator
 from .containerapp_sessionpool_decorator import SessionPoolPreviewDecorator, SessionPoolCreateDecorator, SessionPoolUpdateDecorator
+from .dotnet_component_decorator import DotNetComponentDecorator
 from ._client_factory import handle_raw_exception, handle_non_404_status_code_exception
 from ._clients import (
     GitHubActionPreviewClient,
@@ -94,7 +97,8 @@ from ._clients import (
     ConnectedEnvStorageClient,
     ConnectedEnvCertificateClient,
     JavaComponentPreviewClient,
-    SessionPoolPreviewClient
+    SessionPoolPreviewClient,
+    DotNetComponentPreviewClient
 )
 from ._dev_service_utils import DevServiceUtils
 from ._models import (
@@ -2359,7 +2363,7 @@ def set_environment_telemetry_data_dog(cmd,
 
     containerapp_env_telemetry_data_dog_decorator.construct_payload()
     r = containerapp_env_telemetry_data_dog_decorator.update()
-    
+
     return r
 
 
@@ -2382,7 +2386,7 @@ def delete_environment_telemetry_data_dog(cmd,
 
     containerapp_env_telemetry_data_dog_decorator.construct_payload()
     r = containerapp_env_telemetry_data_dog_decorator.update()
-    
+
     return r
 
 
@@ -2397,7 +2401,7 @@ def show_environment_telemetry_data_dog(cmd,
         raw_parameters=raw_parameters,
         models=CONTAINER_APPS_SDK_MODELS
     )
-    
+
     containerapp_env_def = None
     try:
         containerapp_env_def = containerapp_env_telemetry_data_dog_decorator.show()
@@ -2429,7 +2433,7 @@ def set_environment_telemetry_app_insights(cmd,
 
     containerapp_env_telemetry_app_insights_decorator.construct_payload()
     r = containerapp_env_telemetry_app_insights_decorator.update()
-    
+
     return r
 
 
@@ -2451,7 +2455,7 @@ def delete_environment_telemetry_app_insights(cmd,
 
     containerapp_env_telemetry_app_insights_decorator.construct_payload()
     r = containerapp_env_telemetry_app_insights_decorator.update()
-    
+
     return r
 
 
@@ -2466,7 +2470,7 @@ def show_environment_telemetry_app_insights(cmd,
         raw_parameters=raw_parameters,
         models=CONTAINER_APPS_SDK_MODELS
     )
-    
+
     containerapp_env_def = None
     try:
         containerapp_env_def = containerapp_env_telemetry_app_insights_decorator.show()
@@ -2520,7 +2524,7 @@ def add_environment_telemetry_otlp(cmd,
 
     containerapp_env_telemetry_otlp_decorator.construct_payload()
     r = containerapp_env_telemetry_otlp_decorator.update()
-    
+
     return r
 
 
@@ -2560,7 +2564,7 @@ def update_environment_telemetry_otlp(cmd,
 
     containerapp_env_telemetry_otlp_decorator.construct_payload()
     r = containerapp_env_telemetry_otlp_decorator.update()
-    
+
     return r
 
 
@@ -2583,7 +2587,7 @@ def remove_environment_telemetry_otlp(cmd,
 
     containerapp_env_telemetry_otlp_decorator.construct_remove_payload()
     r = containerapp_env_telemetry_otlp_decorator.update()
-    
+
     return r
 
 
@@ -2599,7 +2603,7 @@ def show_environment_telemetry_otlp(cmd,
         raw_parameters=raw_parameters,
         models=CONTAINER_APPS_SDK_MODELS
     )
-    
+
     containerapp_env_def = None
     try:
         containerapp_env_def = containerapp_env_telemetry_otlp_decorator.show()
@@ -2615,7 +2619,7 @@ def show_environment_telemetry_otlp(cmd,
 
         if not otlp:
             raise ResourceNotFoundError(f"Otlp entry with name --otlp-name {otlp_name} does not exist, please retry with different name")
-        
+
         existing_otlps = otlp
         safe_set(containerapp_env_def, "properties", "openTelemetryConfiguration", "destinationsConfiguration", "otlpConfigurations", value=existing_otlps)
 
@@ -2633,7 +2637,7 @@ def list_environment_telemetry_otlp(cmd,
         raw_parameters=raw_parameters,
         models=CONTAINER_APPS_SDK_MODELS
     )
-    
+
     containerapp_env_def = None
     try:
         containerapp_env_def = containerapp_env_telemetry_otlp_decorator.show()
@@ -2759,3 +2763,65 @@ def delete_session_pool(cmd,
 
     return r
 
+
+def list_dotnet_components(cmd, environment_name, resource_group_name):
+    raw_parameters = locals()
+    dotnet_component_decorator = DotNetComponentDecorator(
+        cmd=cmd,
+        client=DotNetComponentPreviewClient,
+        raw_parameters=raw_parameters,
+        models=CONTAINER_APPS_SDK_MODELS
+    )
+    return dotnet_component_decorator.list()
+
+
+def show_dotnet_component(cmd, dotnet_component_name, environment_name, resource_group_name):
+    raw_parameters = locals()
+    dotnet_component_decorator = DotNetComponentDecorator(
+        cmd=cmd,
+        client=DotNetComponentPreviewClient,
+        raw_parameters=raw_parameters,
+        models=CONTAINER_APPS_SDK_MODELS
+    )
+    result = dotnet_component_decorator.show()
+    if result is not None:
+        logger.warning("Found DotNet Component '%s' in environment '%s' in resource group '%s'.", dotnet_component_name, environment_name, resource_group_name)
+
+    component_type = safe_get(result, "properties", "componentType")
+    if component_type == DOTNET_COMPONENT_RESOURCE_TYPE:
+        aspire_dashboard_url = dotnet_component_decorator._get_aspire_dashboard_url(environment_name, resource_group_name, dotnet_component_name)
+        logger.warning("Aspire Dashboard URL: %s.", aspire_dashboard_url)
+    return result
+
+
+def delete_dotnet_component(cmd, dotnet_component_name, environment_name, resource_group_name, no_wait=False):
+    raw_parameters = locals()
+    dotnet_component_decorator = DotNetComponentDecorator(
+        cmd=cmd,
+        client=DotNetComponentPreviewClient,
+        raw_parameters=raw_parameters,
+        models=CONTAINER_APPS_SDK_MODELS
+    )
+    logger.warning("Deleting DotNet Component '%s' in environment '%s' in resource group '%s'.", dotnet_component_name, environment_name, resource_group_name)
+    dotnet_component_decorator.delete()
+    logger.warning("Successfully deleted DotNet Component '%s' in environment '%s' in resource group '%s'.", dotnet_component_name, environment_name, resource_group_name)
+
+
+def create_dotnet_component(cmd, dotnet_component_name, environment_name, resource_group_name, dotnet_component_type=DOTNET_COMPONENT_RESOURCE_TYPE, no_wait=False):
+    raw_parameters = locals()
+    dotnet_component_decorator = DotNetComponentDecorator(
+        cmd=cmd,
+        client=DotNetComponentPreviewClient,
+        raw_parameters=raw_parameters,
+        models=CONTAINER_APPS_SDK_MODELS
+    )
+    dotnet_component_decorator.construct_payload()
+    logger.warning("Creating DotNet Component '%s' in environment '%s' in resource group '%s.", dotnet_component_name, environment_name, resource_group_name)
+    r = dotnet_component_decorator.create()
+    if r is not None:
+        logger.warning("Successfully created DotNet Component '%s' in environment '%s' in resource group '%s'.", dotnet_component_name, environment_name, resource_group_name)
+    component_type = safe_get(r, "properties", "componentType")
+    if component_type == DOTNET_COMPONENT_RESOURCE_TYPE:
+        aspire_dashboard_url = dotnet_component_decorator._get_aspire_dashboard_url(environment_name, resource_group_name, dotnet_component_name)
+        logger.warning("Access your Aspire Dashboard at %s.", aspire_dashboard_url)
+    return